test/integration_test/requirements.txt
asn1crypto==1.5.1
attrs==25.1.0
autopep8==2.3.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coincurve==20.0.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pathspec==0.12.1
pluggy==1.5.0
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pyparsing==3.2.1
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3
safe-pysha3==1.0.4
urllib3==2.3.0
wheel==0.45.1